Created ProjectValidationItems status class
