+++ Editor/ProjectValidation/Items/ProjectValidationItems.cs
@@ -2,9 +2,7 @@ using UnityEngine;
 using Cognitive3D.Components;
 using UnityEditor;
 using UnityEngine.SceneManagement;
-using UnityEditor.SceneManagement;
 using System.Threading.Tasks;
-using System.Collections.Generic;
 
 namespace Cognitive3D
 {
@@ -13,41 +11,14 @@ namespace Cognitive3D
     {
         private const ProjectValidation.ItemCategory CATEGORY = ProjectValidation.ItemCategory.All;
         private const int INITIAL_DELAY_IN_SECONDS = 1;
-        internal static Dictionary<string, bool> sceneVaidationStatusDic = new Dictionary<string, bool>();
 
         static ProjectValidationItems()
         {
             WaitBeforeProjectValidation();
-            EditorSceneManager.sceneOpened += OnSceneOpened;
-
-            EditorSceneManager.sceneSaved += OnSceneSaved;
-            EditorSceneManager.sceneClosed += OnSceneClosed;
-        }
-
-        private static void OnSceneClosed(Scene scene)
-        {
-            AddOrUpdateSceneValidationStatus(scene);
-        }
-
-        private static void OnSceneSaved(Scene scene)
-        {
-            AddOrUpdateSceneValidationStatus(scene);
-        }
-
-        static void AddOrUpdateSceneValidationStatus(Scene scene)
-        {
-            if (!sceneVaidationStatusDic.ContainsKey(scene.name))
-            {
-                sceneVaidationStatusDic.Add(scene.name, ProjectValidation.hasNotFixedItems());
-            }
-            else
-            {
-                sceneVaidationStatusDic[scene.name] = ProjectValidation.hasNotFixedItems();
-            }
         }
 
         // Adding a delay before adding and verifying items to ensure the scene is completely loaded in the editor
-        static async void WaitBeforeProjectValidation()
+        internal static async void WaitBeforeProjectValidation()
         {
             await Task.Delay(INITIAL_DELAY_IN_SECONDS * 1000);
 
@@ -56,18 +27,9 @@ namespace Cognitive3D
             ProjectValidationGUI.Reset();
         }
 
-        // Update project validation items when a new scene opens
-        private static void OnSceneOpened(Scene scene, OpenSceneMode mode)
-        {
-            Debug.Log("@@@ Scene is opened!");
-            ProjectValidation.Reset();
-            WaitBeforeProjectValidation();
-        }
-
         private static void AddProjectValidationItems()
         {
             // Required Items
-            // TODO: Is the fix action good?
             ProjectValidation.AddItem(
                 level: ProjectValidation.ItemLevel.Required, 
                 category: CATEGORY,
@@ -160,7 +122,6 @@ namespace Cognitive3D
                 }
                 );
 
-            // Fix action?
             ProjectValidation.AddItem(
                 level: ProjectValidation.ItemLevel.Required, 
                 category: CATEGORY,
