+++ Editor/ProjectValidation/Items/ProjectValidationItemsStatus.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0eb88f9e865501a44b1d76775b609b1d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
